Create General use case components
Create components that can be used for general use case.

The navigation bar will probably be reused by more
than 3 pages. I think it will create problems with
implementing the shadow to indicate current page inside
the nav bar.

I want to reuse the standard head and body layout
for most pages. So that the imports for the links
and scripts can be placed in one file only.