Controller for operations on Account & Orders
The Orders controller is currently on progress and not
yet finished.

The Account controller is stable as of writing this
message.

The error messages from the validation process is
not handled on the client side.

I don't know
how to process the automatic response given by
the laravel request form-data-validation;

Any error handled by the controllers will automatically
reload the current page.