Form Request Checker for Login & Register
The request is created using the artisan command.

`php artisan make:request nameOfRequestHandler`

I like the builtin request data filter(idk the actual
name of the value checker).

I did not create custom rules, I only used the built-in rules.

I wanted to create a custom message for every error that occurs.

So I made use of custom request to manage the message and rules
away from the code of the controller.

I like the built-in rules I don't have to create code
to check the values of the form data.

Now I don't know how to manage the error messages from the custom request.