Fix: Order Column change to match Categories table
To match the column value and data of the Categories
table, the incorrect name and data-type of the
Order column "category" was changed to the
datatype of integer and the name is changed to
"category_id".

I did not properly read the pdf about the project
requirements so I made a mistake regarding the column
names and datatypes.

I kind of did not read the pdf and lessons which is
why I made a lot of mistakes with regards to managing
the data using eloquent.