Added core files of bootstrap and jquery to public
To use bootstrap 5 and jquery, I copied the
minified files from their dist folder to public folder
of the project.

I don't want to deal with additional set-up when I just wanted to
use the packages for design and utilities.

And I kind of got short on time so I took the shortcut,
instead of setting up the config.