Working Base Log-in & Register forms
The basic structure for the log-in and
register of accounts.

The pages do not use bootstrap and jquery because when I was writing the
code, I haven't fix the problems with regards
to external package dependencies.

I did not implement bacic css styling and scripting
because I just wanted it to work as fast as possible.