Orders Page Content display All Table Data
The orders view originally should retreive the data from the
database to pass to the read view to show all the orders data.

But I encountered an error when trying out reading data from
inside a view.

There is a solution in stackoverflow, but is a bit complicated.

I chose the simpler method of leaving the data retrieval to the
controller.

The controller will pass the data to the view.

Which leaves orders page view in charge of extending the
orders table read view.