Workng Create view & Pending Update view
The Create view for the orders is working fine.

Aside from error handling. Entry creation is visible on
the client side through the read view.

The update view is currently on progress.

The update and create view almost have the same structure,
though they both differ in parameters.

The update view will present default values and placeholder data
for the update form unlike the create view.