Fix: Update & Delete no longer use Ajax
Using ajax when the page needs to reload is a mistake
in my part.

I want the page to reload after the request, but
since ajax was used to create the request
the page was not reloading.

Visited stackoverflow for advice on the problem.

Got Recommended to not use ajax if page will reload
and the fix I chose is to change the window location
with a url combined with the GET data.

Not secure but works fast for development.