Wrong set-up for Laravel debugbar.
    console.log("entry id : " + entryId);
Tried to follow the tutorial for setting-up laravel debugbar
but there are some differences with the project configuration
shown and the project config of my project.

I tried to follow on and refer the service provider
by laravel debugbar but I got an error instead.

So I tried running the Laravel Debugbar functions,
the info function, and it works without the
additional config tweaking.

I first commented my changes, but I found it too dirty so
I just removed it all together.