Fix Rule & Added functions for Delete & Update
Added a blank rule just to add the id data for
the update, and in case of a creation operation,
if the id data is blank, then it will be
ignored since it does not have any rules associated
with it.

Added functions for the update and delete operations.

The update function takes the parameters from an incoming
get request using the default Request class,
I did not run any validation for the Id variable.

The validation checking for the other fields is the
same for the create process. Both create and update
use the same validation rules.

The Delete function is a bit risky. No validation
and checks, if the id matches, then it is removed
straigth away.

The update function is messy with comments because
I debugged it a lot to figure out the problem
why it won't read the data from the table.